prints the first 50 Fibonacci numbers, starting with 1 and 2

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
  * Return: The first 50 fib numbers
  */
 
-int fib(int n)
+int fib(long int n)
 {
 	if (n <= 1)
 		return (n);
